requirements.txt
cmaes==0.12.0
matplotlib==3.10.7
numpy==2.3.5
opencv_python==4.11.0.86
pandas==2.3.3
panel==1.8.4
param==2.3.1
Pillow==12.0.0
segmentation==0.2.2
torch==2.8.0
tqdm==4.67.1
~orch==2.7.1
